Remove CurseForge from the download options.
CurseForge moderation considers that "we [CurseForge] are afraid that
other users may use your mod in a lewd or inappropriate manner due to
how the animation looks like".
And with my support ticket being inactive for a week now, I don't
believe that the mod will get any time soon on that platform.